+++ globalConfig.json
@@ -1,1 +1,1 @@
-{"mongoUri":"mongodb://127.0.0.1:51464/jest?","mongoDBName":"jest"}
+{"mongoUri":"mongodb://127.0.0.1:51508/jest?","mongoDBName":"jest"}
