+++ src/main/routes/signup-routes.ts
@@ -1,7 +1,7 @@
 /* eslint-disable @typescript-eslint/no-misused-promises */
 import { Router } from 'express'
 import { makeSignUpController } from '../factories/signup/signup-factory'
-import { adaptRoute } from '../adapters/express-route-adapter'
+import { adaptRoute } from '../adapters/express/express-route-adapter'
 
 export default (router: Router): void => {
   router.post('/signup', adaptRoute(makeSignUpController()))
